requirements.txt
pandas>=2.1,<3.0
PyYAML>=6.0,<7.0
tqdm>=4.66,<5.0
matplotlib>=3.8,<4.0
opencv-python-headless>=4.10,<5.0
ultralytics>=8.3.0,<9.0
